@@ -161,12 +162,12 @@
     "cfonts": "^2.8.6",
     "chai": "^4.2.0",
     "chalk": "^2.4.2",
-    "copy-webpack-plugin": "^6.1.0",
+    "copy-webpack-plugin": "^6.1.1",
     "cross-env": "^5.2.1",
     "css-loader": "^2.1.1",
     "del": "^4.1.1",
     "devtron": "^1.4.0",
-    "electron": "^9.1.2",
+    "electron": "^9.3.1",
     "electron-builder": "^22.8.1",
     "electron-toolkit": "^1.0.24",
     "file-loader": "^3.0.1",
