@@ -1,17 +1,17 @@
-import { Zcoind } from '../zcoind';
+import { Firod } from '../firod';
 
 import { createLogger } from '../../lib/logger';
-const logger = createLogger('zcoin:daemon:settings');
+const logger = createLogger('firo:daemon:settings');
 
-export function handleEvent(store, zcoin: Zcoind, eventData: any) {
+export function handleEvent(store, firo: Firod, eventData: any) {
     let key = Object.keys(eventData)[0];
     let mn = eventData[key];
     store.dispatch('Masternode/updateMasternode', mn);
 }
 
-export async function initialize(store: any, zcoind: Zcoind) {
-    await zcoind.awaitApiIsReady();
+export async function initialize(store: any, firod: Firod) {
+    await firod.awaitApiIsReady();
 
-    const data = await zcoind.getMasternodeList();
+    const data = await firod.getMasternodeList();
     store.dispatch('Masternode/updateMasternodeList', Object.values(data));
 }
