@@ -1,8 +1,8 @@
-import { Zcoind } from '../zcoind';
+import { Firod } from '../firod';
 
 import { createLogger } from '../../lib/logger';
-const logger = createLogger('zcoin:daemon:settings');
+const logger = createLogger('firo:daemon:settings');
 
-export function handleEvent(store, zcoin: Zcoind, eventData: any) {
+export function handleEvent(store, firo: Firod, eventData: any) {
     store.dispatch('Masternode/updateMasternodeList', Object.values(eventData));
 }
