@@ -1,11 +1,11 @@
-import { Zcoind } from '../zcoind';
+import { Firod } from '../firod';
 
 import { createLogger } from '../../lib/logger';
-const logger = createLogger('zcoin:daemon:paymentRequest');
+const logger = createLogger('firo:daemon:paymentRequest');
 
 
-export async function initialize(store: any, zcoind: Zcoind) {
-    const data = await zcoind.send(null, 'initial', 'paymentRequest', {});
+export async function initialize(store: any, firod: Firod) {
+    const data = await firod.send(null, 'initial', 'paymentRequest', {});
     logger.info("Got initial paymentRequest data");
     store.commit('PaymentRequest/setStateWithInitialPaymentRequest', data);
 }
