@@ -1,12 +1,12 @@
-import {Zcoind} from "../zcoind";
+import {Firod} from "../firod";
 
 import { createLogger } from '../../lib/logger';
-const logger = createLogger('zcoin:daemon:settings');
+const logger = createLogger('firo:daemon:settings');
 
-export async function initialize(store: any, zcoind: Zcoind) {
-    await zcoind.awaitApiIsReady();
+export async function initialize(store: any, firod: Firod) {
+    await firod.awaitApiIsReady();
 
-    const data = await zcoind.getSettings();
+    const data = await firod.getSettings();
     logger.info("Got initial settings data");
     store.commit('Settings/setDaemonSettings', data);
 }
