@@ -1,5 +1,5 @@
-import { Zcoind } from '../zcoind';
+import { Firod } from '../firod';
 
-export function handleEvent(store, zcoin: Zcoind, eventData: any) {
+export function handleEvent(store, firo: Firod, eventData: any) {
     store.dispatch('Transactions/handleTransactionEvent', eventData);
 }
