@@ -1,5 +1,7 @@
 <template>
     <div id="app">
+        <div id="app-drag-area" />
+
         <div :v-show="!!waitingReason">
             <WaitingScreen v-if="waitingReason" :reason="waitingReason" />
         </div>
@@ -17,7 +19,7 @@ import { mapGetters } from 'vuex'
 import WaitingScreen from "renderer/components/WaitingScreen";
 
 export default {
-    name: 'ZcoinClient',
+    name: 'FiroClient',
 
     components: {
         WaitingScreen
@@ -29,7 +31,24 @@ export default {
 }
 </script>
 
-<style>
-@import '../../node_modules/lato-font/css/lato-font.css';
-@import '../../node_modules/typeface-overpass-mono/index.css';
+<style lang="scss">
+@import './styles';
+@import './styles/z';
+
+#app {
+    position: fixed;
+    user-select: none;
+
+    #app-drag-area {
+        position: absolute;
+        top: 0;
+        z-index: $z-app-drag-area;
+        height: $size-app-drag-area-height;
+        width: 100vw;
+
+        // Allow the user to drag the window from this area.
+        user-select: none;
+        -webkit-app-region: drag;
+    }
+}
 </style>
