@@ -1,108 +1,297 @@
 <template>
-    <section class="tx-page">
-        <div class="window-height payment-page-window">
-            <section
-                v-scrollable
-                class="payment-list-container"
-            >
-                <payments-list />
-            </section>
-            <section
-                v-if="isShowingCustomInputs"
-                class="overlay centered"
-            >
-                <custom-input-popup />
-            </section>
+    <section class="payments-page">
+        <div class="inner">
+            <div class="top-section">
+                <input
+                    v-model="filter"
+                    type="text"
+                    class="table-filter-input"
+                    :placeholder="$t('send.table__outgoing-payments.placeholder__filter')"
+                />
 
-            <section
-                v-if="openAddressBook != null && openAddressBook.open"
-                class="overlay centered"
-            >
-                <address-book-popup />
-            </section>
+                <div v-if="showUnsyncedWarning" class="show-unsynced-warning">
+                    The blockchain is not yet synced. Payment information may be incomplete or inaccurate.
+                </div>
 
-            <WarningWalletWithoutMnemonics v-if="apiStatus.data && !apiStatus.data.hasMnemonic && apiStatus.data.shouldShowWarning && showWarning" @close-mnemonic-warning="closeMnemonicWarning"/>
+                <div v-if="newTableData.length" class="awaiting-updates">
+                    New payments have arrived. <a href='#' @click="reloadTable">Click here</a> to load new transactions.
+                </div>
+            </div>
+
+            <animated-table
+                :data="filteredTableData"
+                :fields="tableFields"
+                track-by="id"
+                :no-data-message="tableData.length ? 'No transactions matched your search criterion' : 'No Payments made yet.'"
+                :sort-order="sortOrder"
+                :compare-elements="comparePayments"
+                :per-page="17"
+                :on-page-change="(pageNumber) => this.currentPage = pageNumber"
+            />
         </div>
-
-        <section class="tx-page-sidebar">
-            <router-view />
-        </section>
     </section>
 </template>
 
-
 <script>
 import { mapGetters } from 'vuex';
-import PaymentsList from 'renderer/components/PaymentsList';
-import CustomInputPopup from 'renderer/components/Overlay/CustomInputPopup';
-import AddressBookPopup from 'renderer/components/Overlay/AddressBookPopup';
-import WarningWalletWithoutMnemonics from './WarningWalletWithoutMnemonics.vue';
+
+import AnimatedTable from 'renderer/components/AnimatedTable/AnimatedTable';
+import RelativeDate from 'renderer/components/AnimatedTable/AnimatedTableRelativeDate';
+import Amount from 'renderer/components/AnimatedTable/AnimatedTableAmount';
+import Label from 'renderer/components/AnimatedTable/AnimatedTableLabel';
+
+import { convertToCoin } from "lib/convert";
+
+const tableFields = [
+    {name: RelativeDate},
+    {name: Label},
+    {name: Amount}
+];
 
 export default {
-    name: 'PaymentsPage',
+    name: 'PaymentsList',
 
     components: {
-        PaymentsList,
-        CustomInputPopup,
-        WarningWalletWithoutMnemonics,
-        AddressBookPopup
+        AnimatedTable,
     },
-    data() {
-        return {
-            data: ''
+
+    props: {
+        selectedPayment: {
+            type: String,
+            default: null
         }
     },
-    mounted() {
-        this.$on('close-mnemonic-warning', this.closeMnemonicWarning);
+
+    data () {
+        return {
+            tableFields,
+            filter: '',
+            tableData: [],
+            newTableData: [],
+            currentPage: 1
+        }
     },
+
+    watch: {
+        latestTableData: {
+            immediate: true,
+
+            handler() {
+                this.newTableData = this.latestTableData;
+                if (this.currentPage === 1) {
+                    this.reloadTable();
+                }
+            }
+        }
+    },
+
     computed: {
         ...mapGetters({
-            isShowingCustomInputs: 'ZcoinPayment/customInputs',
-            hasApiStatus: 'ApiStatus/hasApiStatus',
-            apiStatus: 'ApiStatus/apiStatus',
-            showWarning: 'Settings/showWarning',
-            openAddressBook: 'App/openAddressBook'
+            transactions: 'Transactions/transactions',
+            addresses: 'Transactions/addresses',
+            consolidatedMints: 'Transactions/consolidatedMints',
+            paymentRequests: 'PaymentRequest/paymentRequests',
+            isBlockchainSynced: 'Blockchain/isBlockchainSynced',
+            isReindexing: 'ApiStatus/isReindexing'
         }),
+
+        showUnsyncedWarning() {
+            return !this.isBlockchainSynced || this.isReindexing;
+        },
+
+        latestTableData () {
+            const tableData = [];
+
+            for (const [id, tx] of Object.entries(this.transactions)) {
+                // Mined transactions are incorrectly marked as change.
+                if (tx.isChange && tx.category !== 'mined') {
+                    continue;
+                }
+                // Mints are handled separately.
+                if (tx.category === 'mint') {
+                    continue;
+                }
+
+                if (!['mined', 'receive', 'spendIn', 'send', 'spendOut', 'znode'].includes(tx.category)) {
+                    this.$log.error(`unknown category '${tx.category}' on tx ${id}`);
+                    continue;
+                }
+
+                if (!tx.address) {
+                    this.$log.silly(`transaction ${id} with no associated address`);
+                    continue;
+                }
+
+                // Coordinate this with the default values in AnimatedTableLabel.
+                let extraSearchText;
+                switch (tx.category) {
+                case 'mined':
+                    extraSearchText = 'Mined Transaction';
+                    break;
+
+                case 'receive':
+                case 'spendIn':
+                    extraSearchText = 'Incoming Transaction';
+                    break;
+
+                case 'send':
+                case 'spendOut':
+                    extraSearchText = 'Outgoing Transaction';
+                    break;
+
+                case 'znode':
+                    extraSearchText = 'Znode Payment';
+                    break
+                }
+
+                tableData.push({
+                    // id is the path of the detail route for the transaction.
+                    id: `/transaction-info/${id}`,
+                    txid: tx.txid,
+                    category: tx.category,
+                    blockHeight: tx.blockHeight,
+                    date: tx.blockTime * 1000 || Infinity,
+                    amount: tx.amount,
+                    address: tx.address,
+                    label:
+                        tx.label ||
+                        (this.paymentRequests[tx.address] ? this.paymentRequests[tx.address].label : undefined),
+                    extraSearchText: extraSearchText + (['send', 'spendOut'].includes(tx.category) ? '-' : '+') + convertToCoin(tx.amount)
+                });
+            }
+
+            for (const [blockHeight, mintInfo] of Object.entries(this.consolidatedMints)) {
+                tableData.push({
+                    id: `/mint-info/${blockHeight}`,
+                    category: 'mint',
+                    blockHeight: blockHeight,
+                    date: mintInfo.blockTime * 1000 || Infinity,
+                    amount: mintInfo.totalMintAmount,
+                    label: null,
+                    // Coordinate this with the default values in AnimatedTableLabel.
+                    extraSearchText: 'Private Mint (' + convertToCoin(mintInfo.totalMintAmount) + ')'
+                });
+            }
+
+            for (const pr of Object.values(this.paymentRequests)) {
+                if (this.addresses[pr.address] && this.addresses[pr.address].length) {
+                    // There are actual transactions associated with the request now, so we don't need to show it.
+                    continue;
+                }
+
+                if (pr.state !== 'active') {
+                    // Don't show deleted or archived payment requests.
+                    continue;
+                }
+
+                tableData.push({
+                    // id is the path of the detail route for the payment request.
+                    id: `/payment-request/${pr.address}`,
+                    category: 'payment-request',
+                    blockHeight: null,
+                    date: pr.createdAt,
+                    amount: pr.amount,
+                    label: pr.label,
+                    // Coordinate this with the default values in AnimatedTableLabel.
+                    extraSearchText: 'Payment Request (' + convertToCoin(pr.amount) + ')'
+                });
+            }
+
+            return tableData;
+        },
+
+        filteredTableData () {
+            if (!this.filter) {
+                return this.tableData;
+            }
+
+            let filter = this.filter.toLowerCase();
+            return this.tableData.filter(tableRow =>
+                ['label', 'address', 'category', 'extraSearchText'].find(key =>
+                    tableRow[key] && tableRow[key].toLowerCase().indexOf(filter) !== -1
+                )
+            )
+        },
+
+        sortOrder () {
+            return [
+                {
+                    sortField: 'date',
+                    direction: 'desc'
+                }
+            ]
+        }
     },
 
     methods: {
-        async closeMnemonicWarning() {
-            this.$store.commit('Settings/MNEMONIC_WARNING_SETTING', false);
+        comparePayments(a, b) {
+            return !['id', 'category', 'blockHeight', 'date', 'amount', 'address', 'label'].find(field =>
+                a[field] !== b[field]
+            );
+        },
+
+        reloadTable() {
+            this.tableData = this.newTableData;
+            this.newTableData = [];
         }
     }
 }
 </script>
 
 <style lang="scss" scoped>
-    .tx-page {
-        display: grid;
-        box-sizing: border-box;
-        grid-template-columns: 1fr $detail-view--min-width;
+@import "src/renderer/styles/inputs";
+@import "src/renderer/styles/sizes";
+@import "src/renderer/styles/colors";
 
-        .scrollable {
-            position: relative;
-            z-index: 1;
-            box-sizing: border-box;
-            overflow: scroll;
-            height: 100vh;
-        }
-    }
+.payments-page {
+    height: 100%;
 
-    .payment-list-container,
-    .tx-page-sidebar {
-        position: relative;
-    }
-
-    .payment-list-container {
-        padding: emRhythm(5) emRhythm(4);
+    .inner {
         box-sizing: border-box;
         height: 100%;
-    }
+        padding: $size-main-margin;
 
-    .tx-page-sidebar {
-        background: $color--white;
+        display: flex;
+        flex-flow: column;
+
+        .top-section {
+            height: fit-content;
+
+            .table-filter-input {
+                margin: {
+                    left: auto;
+                    right: 0;
+                    bottom: $size-medium-space;
+                }
+
+                @include wide-rounded-input();
+                max-width: 100%;
+            }
+
+            .filter-input {
+                position: relative;
+                display: inline-block;
+            }
+
+            .show-unsynced-warning, .awaiting-updates {
+                text-align: center;
+                font: {
+                    size: 0.9em;
+                    style: italic;
+                    weight: bold;
+                }
+
+                margin-bottom: 1em;
+            }
+
+            .show-unsynced-warning {
+                color: red;
+            }
+        }
+
+        .animated-table {
+            flex-grow: 1;
+        }
     }
-    .payment-page-window{
-        position: relative
-    }
+}
 </style>
