@@ -3,18 +3,43 @@ import { ipcMain } from 'electron'
 
 import Vue from 'vue'
 import Vuex from 'vuex'
-import modules from './modules'
+
+import AddressBook from "store/modules/AddressBook";
+import AddressValidation from "store/modules/AddressValidation";
+import ApiStatus from "store/modules/ApiStatus";
+import App from "store/modules/App";
+import Balance from "store/modules/Balance";
+import Blockchain from "store/modules/Blockchain";
+import Masternode from "store/modules/Masternode";
+import Notification from "store/modules/Notification";
+import PaymentRequest from "store/modules/PaymentRequest";
+import Settings from "store/modules/Settings";
+import Transactions from "store/modules/Transactions";
+import Window from "store/modules/Window";
 
 import { createLogger } from 'lib/logger'
 
-const logger = createLogger('zcoin:store:main')
+const logger = createLogger('firo:store:main')
 
 Vue.use(Vuex)
 
 const clients = []
 
 const store = new Vuex.Store({
-    modules
+    modules: {
+        AddressBook,
+        AddressValidation,
+        ApiStatus,
+        App,
+        Balance,
+        Blockchain,
+        Masternode,
+        Notification,
+        PaymentRequest,
+        Settings,
+        Transactions,
+        Window
+    }
 })
 
 store.subscribe((mutation, state) => {
