@@ -1,7 +1,7 @@
 import { isEqual } from 'lodash';
 
 const state = {
-    // ApiStatus from src/daemon/zcoind
+    // ApiStatus from src/daemon/firod
     apiStatus: {}
 };
 
@@ -26,6 +26,8 @@ const getters = {
     network: (state) => (state.apiStatus && state.apiStatus.data) ? state.apiStatus.data.network : undefined,
     // Do we have an apiStatus?
     hasApiStatus: (state): boolean => state.apiStatus && !!state.apiStatus.data,
+    // Can our wallet be recovered from a mnemonic?
+    hasMnemonic: (state): boolean => state.apiStatus && state.apiStatus.data && state.apiStatus.data.hasMnemonic,
     // Is the wallet locked? Returns undefined if not yet loaded.
     isLocked: (state): boolean | undefined => (state.apiStatus && state.apiStatus.data) ? state.apiStatus.data.walletLock : undefined,
     isReindexing: (state): boolean => state.apiStatus && state.apiStatus.data && state.apiStatus.data.reindexing,
