@@ -2,20 +2,43 @@ import Vue from 'vue'
 import Vuex from 'vuex'
 import { ipcRenderer } from 'electron'
 
-import modules from 'store/modules'
 import types from 'store/types'
+
+import AddressBook from "store/modules/AddressBook";
+import AddressValidation from "store/modules/AddressValidation";
+import ApiStatus from "store/modules/ApiStatus";
+import App from "store/modules/App";
+import Balance from "store/modules/Balance";
+import Blockchain from "store/modules/Blockchain";
+import Masternode from "store/modules/Masternode";
+import Notification from "store/modules/Notification";
+import PaymentRequest from "store/modules/PaymentRequest";
+import Settings from "store/modules/Settings";
+import Transactions from "store/modules/Transactions";
+import Window from "store/modules/Window";
 
 Vue.use(Vuex)
 
 const localModules = [
-    // todo reduce modules here by looking for local: true in the module definition
-    // Object.keys(modules).map((name) => { modules[name].local ? name : )
     'AppRouter',
     'Router'
 ]
 
 const store = new Vuex.Store({
-    modules,
+    modules: {
+        AddressBook,
+        AddressValidation,
+        ApiStatus,
+        App,
+        Balance,
+        Blockchain,
+        Masternode,
+        Notification,
+        PaymentRequest,
+        Settings,
+        Transactions,
+        Window
+    },
     strict: process.env.NODE_ENV !== 'production' // this up to you
 })
 
